Fix support for GPU selection using CLI argument
