+++ fastchat/model/__init__.py
@@ -1,5 +1,5 @@
+from fastchat.args.add_model_args import add_model_args
 from fastchat.model.model_adapter import (
     load_model,
     get_conversation_template,
-    add_model_args,
 )
